% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somme.R
\name{somme}
\alias{somme}
\title{sum}
\usage{
somme(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
a number
}
\description{
This function simply returns the sum between 2 numbers.
}
\examples{
somme(1,2)
somme(0,1)
}
